Changes related to temperature feature
1. Changed u64 to int to support negative
temperature values, i.e. temp1_lcrit may
be as low as -40C

2. In gui surround temperature reading with
square brackets instead of round brackets to
be consistent with the rest of the drive status line

+++ src/context.h
@@ -128,16 +128,16 @@ typedef struct nwipe_context_t_
     u64 throughput;  // Average throughput in bytes per second.
     u64 verify_errors;  // The number of verification errors across all passes.
     char temp1_path[MAX_HWMON_PATH_LENGTH];  // path to temperature variables /sys/class/hwmon/hwmonX/ etc.
-    u64 temp1_crit;  // Critical high drive temperature, 1000000=unitialised, millidegree celsius.
-    u64 temp1_highest;  // Historical highest temperature reached, 1000000=unitialised, millidegree celsius.
-    u64 temp1_input;  // drive temperature, -1=unitialised. 1000000=unitialised, millidegree celsius.
-    u64 temp1_lcrit;  // Critical low drive temperature, 1000000=unitialised, millidegree celsius.
-    u64 temp1_lowest;  // Historically lowest temperature, 1000000=unitialised, millidegree celsius.
-    u64 temp1_max;  // Maximum allowed temperature, 1000000=unitialised, millidegree celsius.
-    u64 temp1_min;  // Miniumum allowed temperature, 1000000=unitialised, millidegree celsius.
-    u64 temp1_monitored_wipe_max;
-    u64 temp1_monitored_wipe_min;
-    u64 temp1_monitored_wipe_avg;
+    int temp1_crit;  // Critical high drive temperature, 1000000=unitialised, millidegree celsius.
+    int temp1_highest;  // Historical highest temperature reached, 1000000=unitialised, millidegree celsius.
+    int temp1_input;  // drive temperature, -1=unitialised. 1000000=unitialised, millidegree celsius.
+    int temp1_lcrit;  // Critical low drive temperature, 1000000=unitialised, millidegree celsius.
+    int temp1_lowest;  // Historically lowest temperature, 1000000=unitialised, millidegree celsius.
+    int temp1_max;  // Maximum allowed temperature, 1000000=unitialised, millidegree celsius.
+    int temp1_min;  // Miniumum allowed temperature, 1000000=unitialised, millidegree celsius.
+    int temp1_monitored_wipe_max;
+    int temp1_monitored_wipe_min;
+    int temp1_monitored_wipe_avg;
     time_t temp1_time;  // The time when temperature was last checked, seconds since epoch
     int wipe_status;  // Wipe finished = 0, wipe in progress = 1, wipe yet to start = -1.
     int spinner_idx;  // Index into the spinner character array
